ms_cron Simplified Schedule struct design
Inheritance was provided to allow the possibility of creating more
than one implementation of the Schedule. But it creates problems with
interfaces and circular dependencies due to non-inherited fields, and
it is not worth it because a different implementation will most likely
diverge a lot, not only in its Schedule's fields. It will therefore
require implementation as an separate component.

For this reason, we are redesigning this module to be based upon a
single, simple struct, where only the storage layer may be changed.