Cron Do not update a Schedule's UpdateAt field when running it
When a Schedule is run we save that moment as the moment that it was
last executed (Last field). We do not want the Update function to save
that moment as the moment the Schedule was last updated as that is
misleading - the definition of the Schedule has not changed at all.